Make sure event channel doesn't get full
In particular the first analysis often takes a very long time
(~minutes) as lots of nonces need to be checked. During this
time, we receive many new pending transaction events. Those won't
be processed until the analysis is over, so they stay in the
channel. At some point, the channel is full and sending events
blocks. That's a problem for blocks and pools which are sent to
the same channel: The observation timestamp will be quite late.

To fix this, we make sure to only send tx events when there's
plenty of space left in the channel.

At some point, we might also want to increase the channel
capacity.