Treat disappearing transactions as visible
Transactions are considered disappearing in the timespan between a pool
observation that includes the transaction and a subsequent observation that
doesn't. This means that transactions that get included in a block at time t
are usually considered disappearing at time t as well (they leave the pool
once the new block has been observed). As this is normal behavior, we
should treat these transactions as visible.