Broaden replaced tx criterion
Validators are now not expected to include more than one transaction
from each sender. The reason is that the first transaction will change
the balance of the sender, so we cannot be sure that the account would
have been able to pay the fee for the second transaction, even though it
appeared in our mempool.