Store all data in single table
This should increase performance significantly because we avoid LEFT
JOINs.

The only data that is missing from the full_miss table is the number
of total and pool transactions per block. This isn't related to misses
and will likely not be JOINed with it, so it should remain in a separate
table.

Existing data is migrated, but not deleted. For now, both the old set of
tables and the new table are filled. When it's confirmed that it works,
we should delete the old tables and only insert into the new one.