Add txs and blocks API endpoints
These endpoints query the same data as for misses, but group by tx or
block, respectively. They allow to filter by the minimum number of
blocks the tx was missed in or txs that were missed in one block.

The response also contains `from` and `to` timestamps that indicate the
queried range. They can be used for pagination (i.e. use `to` as the
value for the next `from`).

Grouping is done in Rust instead of SQL for simplicity. Besides
performance, there is one issue with that: Misses at the boundary of the
queried interval. A tx that was missed once in the interval, but another
time outside of the interval will not show up in a query with
`min_num_misses=2`. In SQL there are solutions for this, but I didn't
manage to solve this in an efficient manner.