Restructure API
Moved some logic from Rust into SQL, including grouping.

In txs/blocks query, use a JOIN to make sure we got all misses for the
txs. Due to the filters and limit we apply, we might cut off some
misses and as a result get a wrong `num_misses`. This means we get some
misses that are outside the queries time range. To not throw pagination
off the track, keep the timestamp of the "original" miss as ref_time.

Use tx_quorum_reached instead of tx_first_seen as tie breaker when
ordering as we use this as the canonical timestamp of a transaction in
other places as well.