Make time filtering more granular
We used to do time filtering of misses based on the proposal time of the
block, and by extension also of txs and blocks. However, there can be
many misses in a single block and thus with the same proposal. In rare
cases, this can lead to users who paginate through many results to get
stuck or miss some entries.

Now, we filter based on proposal time and optionally tx_quorum_reached
as a secondary field.