Rewrite API
Pagination now with proposal time and offset which should be a good
compromise between efficiency, simplicity, and correctness. Aggregation
is now done in Rust instead of SQL because it's easier to debug and the
process is not trivial. At the moment, the queries return duplicates
which are removed later on in Rust postprocessing which is supoptimal
efficiency-wise, but simpler.