Revert "Ignore cancel transactions during analysis"
This reverts commit 021789ce583c97624486be180b2092d0fbf5a12f.

As it turns out, misses from cancel transactions are not false
positives, but valid. Some builders only replace transactions if the
total priority fee (as opposed to the priority fee per gas) is
increased. This may or may not be a legit strategy. But depending on how
this is implemented, it may result in the replacing transaction to be
fully ignored. Thus, it won't be included even if the original
transaction is not included, leading to worse UX for the user and less
fee revenue for the builder.